Model/requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
sentence-transformers
pinecone-client
nest-asyncio
